+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 
 ## 0.5.0 (unreleased)
 
+- Removed support to load website URL from ``links.website``. From now
+  on, it is expected only in ``links.website.url``.
+
 
 ## 0.4.0 (2024-06-30)